requirements.txt
discord.py==2.3.2
pandas==2.2.1
openpyxl==3.1.2
tqdm==4.66.2
smartsheet-python-sdk==2.178.0
pywin32==306
requests==2.31.0
python-dotenv==1.0.0
setuptools==69.2.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Pillow>=10.0.0